[refactor] `parseFloat` instead of `parseInt` for unit price and total price
Because with `parseInt` if we used something like `0.13` for the unit price, it was considered as a `0`. So it was still doing an error.

+++ app/components/Modal.tsx
@@ -452,7 +452,7 @@ export function Modal({ event, user, examStatus, exams, setExams }: ModalProps) 
                         }
 
                         if (shouldNotifyFinished) {
-                            if(parseInt(boxes) <= 0 || parseInt(priceUnit) <= 0 || parseInt(priceTotal) <= 0) {
+                            if(parseInt(boxes) <= 0 || parseFloat(priceUnit) <= 0 || parseFloat(priceTotal) <= 0) {
                                 window.alert("To change the status to `Finished`, these fields need to be set to something bigger than 0 : Number of boxes, Price Unit, Price Total");
                                 setSelectStatus(previousStatus);
                                 return;
